+++ Realm/RLMSyncConfiguration.mm
@@ -91,6 +91,14 @@ static BOOL isValidRealmURL(NSURL *url) {
         && self.stopPolicy == that.stopPolicy;
 }
 
+- (void)setEnableSSLValidation:(BOOL)enableSSLValidation {
+    _config->client_validate_ssl = (bool)enableSSLValidation;
+}
+
+- (BOOL)enableSSLValidation {
+    return (BOOL)_config->client_validate_ssl;
+}
+
 - (realm::SyncConfig)rawConfiguration {
     return *_config;
 }
@@ -161,6 +169,9 @@ static BOOL isValidRealmURL(NSURL *url) {
             std::move(bindHandler),
             std::move(errorHandler)
         });
+        if (NSNumber *disabled = [[RLMSyncManager sharedManager] globalSSLValidationDisabled]) {
+            _config->client_validate_ssl = ![disabled boolValue];
+        }
         self.customFileURL = customFileURL;
         return self;
     }